Added npm script for removing dev-assets.
This will remove the parcel build cache and the node_modules, none of which need to be committed to the repository.

Technically, I shouldn't even be committing the compiled assets (`.js` and `.css` files); but, I figured that would make it the easiest for anyone who might want to check out this repo and try it out. That way, they don't have to install or build anything - they just have to start the container.